refactor(provider): remove retry logic and simplify chat method implementation
